# api-docs-catalogo.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://192.168.49.2:30083","description":"Generated server url"}],"tags":[{"name":"Serviços","description":"Gerenciamento de serviços"},{"name":"Peças e Insumos","description":"Gerenciamento de peças e insumos"}],"paths":{"/pecas/dar-baixa-estoque":{"put":{"tags":["Peças e Insumos"],"summary":"Realizar baixa de estoque de peças/insumos","operationId":"darBaixaEstoque","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DarBaixaEstoqueRequest"}}}},"required":true},"responses":{"204":{"description":"Baixa de estoque realizada com sucesso"},"400":{"description":"Requisição inválida ou estoque insuficiente"}}}},"/servicos":{"get":{"tags":["Serviços"],"summary":"Listar todos os serviços","operationId":"listarTodos","parameters":[{"name":"arg0","in":"query","required":false,"schema":{"type":"integer","format":"int64","default":0}},{"name":"arg1","in":"query","required":false,"schema":{"type":"integer","format":"int64","default":10}}],"responses":{"200":{"description":"Lista de serviços retornada com sucesso","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["Serviços"],"summary":"Cadastrar novo serviço","operationId":"cadastrar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CadastrarServicoRequest"}}},"required":true},"responses":{"201":{"description":"Serviço cadastrado com sucesso","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Requisição inválida","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/pecas":{"get":{"tags":["Peças e Insumos"],"summary":"Listar todas as peças e insumos","operationId":"listarTodas","parameters":[{"name":"arg0","in":"query","required":false,"schema":{"type":"integer","format":"int64","default":0}},{"name":"arg1","in":"query","required":false,"schema":{"type":"integer","format":"int64","default":10}}],"responses":{"200":{"description":"Lista de peças e insumos retornada com sucesso","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["Peças e Insumos"],"summary":"Cadastrar nova peça ou insumo","operationId":"cadastrar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CadastrarPecaInsumoRequest"}}},"required":true},"responses":{"201":{"description":"Peça/insumo cadastrado com sucesso","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Requisição inválida","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/servicos/{id}":{"get":{"tags":["Serviços"],"summary":"Buscar serviço por ID","operationId":"buscarPorId","parameters":[],"responses":{"200":{"description":"Serviço encontrado","content":{"*/*":{"schema":{"type":"object"}}}},"404":{"description":"Serviço não encontrado","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Serviços"],"summary":"Deletar serviço","operationId":"deletar","parameters":[],"responses":{"204":{"description":"Serviço deletado com sucesso"},"404":{"description":"Serviço não encontrado"}}},"patch":{"tags":["Serviços"],"summary":"Atualizar serviço","operationId":"atualizar","parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AtualizarServicoRequest"}}},"required":true},"responses":{"200":{"description":"Serviço atualizado com sucesso","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Requisição inválida","content":{"*/*":{"schema":{"type":"object"}}}},"404":{"description":"Serviço não encontrado","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/pecas/{id}":{"get":{"tags":["Peças e Insumos"],"summary":"Buscar peça/insumo por ID","operationId":"buscarPorId_1","parameters":[],"responses":{"200":{"description":"Peça/insumo encontrado","content":{"*/*":{"schema":{"type":"object"}}}},"404":{"description":"Peça/insumo não encontrado","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Peças e Insumos"],"summary":"Deletar peça ou insumo","operationId":"deletar_1","parameters":[],"responses":{"204":{"description":"Peça/insumo deletado com sucesso"},"404":{"description":"Peça/insumo não encontrado"}}},"patch":{"tags":["Peças e Insumos"],"summary":"Atualizar peça ou insumo","operationId":"atualizar_1","parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AtualizarPecaInsumoRequest"}}},"required":true},"responses":{"200":{"description":"Peça/insumo atualizado com sucesso","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Requisição inválida","content":{"*/*":{"schema":{"type":"object"}}}},"404":{"description":"Peça/insumo não encontrado","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"DarBaixaEstoqueRequest":{"type":"object","properties":{"itemId":{"type":"string","format":"uuid"},"quantidade":{"type":"integer","format":"int32","minimum":1}},"required":["itemId","quantidade"]},"CadastrarServicoRequest":{"type":"object","properties":{"nome":{"type":"string","minLength":1},"descricao":{"type":"string","minLength":1},"preco":{"type":"number"},"duracaoEstimada":{"type":"integer","format":"int32"},"categoria":{"type":"string","minLength":1}},"required":["categoria","descricao","duracaoEstimada","nome","preco"]},"CadastrarPecaInsumoRequest":{"type":"object","properties":{"nome":{"type":"string","minLength":1},"descricao":{"type":"string","minLength":1},"codigoFabricante":{"type":"string","minLength":1},"marca":{"type":"string","minLength":1},"quantidadeEstoque":{"type":"integer","format":"int32"},"precoCusto":{"type":"number"},"precoVenda":{"type":"number"},"categoria":{"type":"string","minLength":1},"ativo":{"type":"boolean"}},"required":["ativo","categoria","codigoFabricante","descricao","marca","nome","precoCusto","precoVenda"]},"AtualizarServicoRequest":{"type":"object","properties":{"preco":{"type":"number"},"ativo":{"type":"boolean"}},"required":["ativo","preco"]},"AtualizarPecaInsumoRequest":{"type":"object","properties":{"qtd":{"type":"integer","format":"int32"},"precoCusto":{"type":"number"},"precoVenda":{"type":"number"},"ativa":{"type":"boolean"}},"required":["ativa","precoCusto","precoVenda","qtd"]}}}}